requirements.txt
anthropic
groq
flask
sqlalchemy
python-dotenv
gunicorn
